requirements.txt
pandas
scikit-learn
joblib
fastapi
uvicorn
hyperopt
pyarrow
fastparquet
python-dotenv